backend/requirements.txt
fastapi
uvicorn[standard]
requests
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
python-dotenv
python-multipart
